kaiax/valset: Organize
